Conflicts:
	.classpath
	Tests/controller/TestRegisterDeleteUser.java
	src/model/Database.java

+++ src/model/Database.java
@@ -1,16 +1,8 @@
 package model;
 
-import java.io.IOException;
 import java.math.BigDecimal;
 import java.util.LinkedList;
 import java.util.ArrayList;
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.LinkedList;
-import java.util.Map;
-import java.util.Properties;
-import java.util.Scanner;
-
 import javax.servlet.http.HttpSession;
 
 import java.sql.*;
@@ -18,11 +10,14 @@ import java.sql.*;
 public class Database {
 	private Connection connection;
 	private Statement statement;
+	private HttpSession session;
 
-	protected String url = "jdbc:db2://192.86.32.54:5040/DALLASB:retrieveMessagesFromServerOnGetMessage=true;emulateParameterMetaDataForZCalls=1;";
+	private String url = "jdbc:db2://192.86.32.54:5040/DALLASB:retrieveMessagesFromServerOnGetMessage=true;emulateParameterMetaDataForZCalls=1;";
+	private String loginString = "Illegal action";
 
-	public Database() throws ClassNotFoundException, SQLException {
+	public Database(HttpSession session) throws ClassNotFoundException, SQLException {
 		connect();
+		this.session = session;
 	}
 
 	private void connect() throws ClassNotFoundException, SQLException {
@@ -32,25 +27,28 @@ public class Database {
 	}
 
 	public ArrayList<User> searchFor(String keyword) {
-		ArrayList<String> IDs = new ArrayList<>();
-		ArrayList<User> users = new ArrayList<>();
+		if (session != null && session.getAttribute("role").equals("e")) {
+			ArrayList<String> IDs = new ArrayList<>();
+			ArrayList<User> users = new ArrayList<>();
 
-		try {
-			ResultSet resultSet = statement
-					.executeQuery("SELECT * FROM \"DTUGRP05\".\"CUSTOMERS\" WHERE \"CPRNo\" LIKE '%" + keyword
-							+ "%' OR \"Phone\" LIKE '%" + keyword + "%' OR LOWER(\"FullName\") LIKE '%" + keyword
-							+ "%'");
-			while (resultSet.next()) {
-				IDs.add(resultSet.getString("CPRNo"));
+			try {
+				ResultSet resultSet = statement
+						.executeQuery("SELECT * FROM \"DTUGRP05\".\"CUSTOMERS\" WHERE \"CPRNo\" LIKE '%" + keyword
+								+ "%' OR \"Phone\" LIKE '%" + keyword + "%' OR LOWER(\"FullName\") LIKE '%" + keyword
+								+ "%'");
+				while (resultSet.next()) {
+					IDs.add(resultSet.getString("CPRNo"));
+				}
+				resultSet.close();
+			} catch (SQLException e) {
+				e.printStackTrace();
 			}
-			resultSet.close();
-		} catch (SQLException e) {
-			e.printStackTrace();
+			for (String ID : IDs) {
+				users.add(getUser(ID));
+			}
+			return users;
 		}
-		for (String ID : IDs) {
-			users.add(getUser(ID));
-		}
-		return users;
+		return null;
 	}
 
 	public String Login(String cpr, String password) throws SQLException {
@@ -120,74 +118,83 @@ public class Database {
 	}
 
 	public String newAccount(Account account, User user) throws SQLException {
-		String cpr = user.getCPR();
-		String ID = account.getAccountID();
-		String name = account.getName();
-		BigDecimal balance = account.getBalance();
-		BigDecimal interest = account.getInterest();
-		String ISOCode = account.getISOCode();
-		CallableStatement call = connection.prepareCall("{call \"DTUGRP05\".CreateAccount(?, ?, ?, ?, ?, ?, ?) }");
-		call.setString("vCPRNo", cpr);
-		call.setString("vAccID", ID);
-		call.setBigDecimal("vAmount", balance);
-		call.setBigDecimal("vInterest", interest);
-		call.setString("vAccName", name);
-		call.setString("vISOCode", ISOCode);
-		call.registerOutParameter("vOutput", java.sql.Types.VARCHAR);
-		call.execute();
-		return call.getString("vOutput");
-	}
-
-	public String closeAccount(String accountID) throws SQLException {
-		CallableStatement call = connection.prepareCall("{call \"DTUGRP05\".DeleteAccount(?, ?) }");
-		call.setString("vAccID", accountID);
-		call.registerOutParameter("vOutput", java.sql.Types.VARCHAR);
-		call.execute();
-		return call.getString("vOutput");
-	}
-
-	public String editAccount(Account account) throws SQLException {
-		CallableStatement call = connection.prepareCall("{call \"DTUGRP05\".EditAccount(?, ?, ?, ?) }");
-		call.setString("vAccID", account.getAccountID());
-		call.setString("vAccName", account.getName());
-		call.setBigDecimal("vInterest", account.getInterest());
-		call.registerOutParameter("vOutput", java.sql.Types.VARCHAR);
-		call.execute();
-		return call.getString("vOutput");
-	}
-
-	public String processTransaction(String type, String accountID, String accountID2, BigDecimal amount,
-			String ISOCode, String transactionName) throws SQLException {
-		CallableStatement call;
-		switch (type) {
-		case "Deposit":
-			call = connection.prepareCall("{call \"DTUGRP05\".Deposit(?, ?, ?, ?) }");
-			call.setString("vAccID", accountID);
-			call.setBigDecimal("vAmount", amount);
-			call.setString("vISOCode", ISOCode);
-			call.registerOutParameter("vOutput", java.sql.Types.VARCHAR);
-			call.execute();
-			return call.getString("vOutput");
-		case "Withdraw":
-			call = connection.prepareCall("{call \"DTUGRP05\".withdraw(?, ?, ?, ?) }");
-			call.setString("vAccID", accountID);
-			call.setBigDecimal("vAmount", amount);
-			call.setString("vISOCode", ISOCode);
-			call.registerOutParameter("vOutput", java.sql.Types.VARCHAR);
-			call.execute();
-			return call.getString("vOutput");
-		case "Transfer":
-			call = connection.prepareCall("{call \"DTUGRP05\".MoneyTransfer(?, ?, ?, ?, ?, ?) }");
-			call.setBigDecimal("vTransfer", amount);
-			call.setString("vTransName", transactionName);
-			call.setString("vAccID1", accountID);
-			call.setString("vAccID2", accountID2);
+		if (session != null && session.getAttribute("role").equals("e")) {
+			String cpr = user.getCPR();
+			String ID = account.getAccountID();
+			String name = account.getName();
+			BigDecimal balance = account.getBalance();
+			BigDecimal interest = account.getInterest();
+			String ISOCode = account.getISOCode();
+			CallableStatement call = connection.prepareCall("{call \"DTUGRP05\".CreateAccount(?, ?, ?, ?, ?, ?, ?) }");
+			call.setString("vCPRNo", cpr);
+			call.setString("vAccID", ID);
+			call.setBigDecimal("vAmount", balance);
+			call.setBigDecimal("vInterest", interest);
+			call.setString("vAccName", name);
 			call.setString("vISOCode", ISOCode);
 			call.registerOutParameter("vOutput", java.sql.Types.VARCHAR);
 			call.execute();
 			return call.getString("vOutput");
 		}
-		return "";
+		return loginString;
+	}
+
+	public String closeAccount(String accountID) throws SQLException {
+		if (session != null && session.getAttribute("role").equals("e")) {
+			CallableStatement call = connection.prepareCall("{call \"DTUGRP05\".DeleteAccount(?, ?) }");
+			call.setString("vAccID", accountID);
+			call.registerOutParameter("vOutput", java.sql.Types.VARCHAR);
+			call.execute();
+			return call.getString("vOutput");
+		}
+		return loginString;
+	}
+
+	public String editAccount(Account account) throws SQLException {
+		if (session != null && session.getAttribute("role").equals("e")) {
+			CallableStatement call = connection.prepareCall("{call \"DTUGRP05\".EditAccount(?, ?, ?, ?) }");
+			call.setString("vAccID", account.getAccountID());
+			call.setString("vAccName", account.getName());
+			call.setBigDecimal("vInterest", account.getInterest());
+			call.registerOutParameter("vOutput", java.sql.Types.VARCHAR);
+			call.execute();
+			return call.getString("vOutput");
+		}
+		return loginString;
+	}
+
+	public String processTransaction(String type, String accountID, String accountID2, BigDecimal amount,
+			String ISOCode, String transactionName) throws SQLException {
+		System.out.println(session);
+		if (session != null && session.getAttribute("loggedinuser") != null) {
+			CallableStatement call;
+			switch (type) {
+			case "Deposit":
+				call = connection.prepareCall("{call \"DTUGRP05\".Deposit(?, ?, ?, ?) }");
+				call.setString("vAccID", accountID);
+				call.setBigDecimal("vAmount", amount);
+				call.setString("vISOCode", ISOCode);
+				call.registerOutParameter("vOutput", java.sql.Types.VARCHAR);
+				call.execute();
+				return call.getString("vOutput");
+			case "Withdraw":
+				call = connection.prepareCall("{call \"DTUGRP05\".withdraw(?, ?, ?, ?) }");
+				call.setString("vAccID", accountID);
+				call.setBigDecimal("vAmount", amount);
+				call.setString("vISOCode", ISOCode);
+				call.registerOutParameter("vOutput", java.sql.Types.VARCHAR);
+				call.execute();
+				return call.getString("vOutput");
+			case "Transfer":
+				call = connection.prepareCall("{call \"DTUGRP05\".MoneyTransfer(?, ?, ?, ?, ?, ?) }");
+				call.setBigDecimal("vTransfer", amount);
+				call.setString("vTransName", transactionName);
+				call.setString("vAccID1", accountID);
+				call.setString("vAccID2", accountID2);
+				return call.getString("vOutput");
+			}
+		}
+		return loginString;
 	}
 
 	public LinkedList<User> getOwners(String accountID) {
@@ -207,11 +214,11 @@ public class Database {
 		return owners;
 	}
 
-	public String getCity(String zipcode) {
+	public String getCity(String postcode) {
 		String city = null;
 		try {
 			ResultSet resultset = statement
-					.executeQuery("select * from \"DTUGRP05\".\"CITIES\" WHERE \"Postcode\" = '" + zipcode + "'");
+					.executeQuery("select * from \"DTUGRP05\".\"CITIES\" WHERE \"Postcode\" = '" + postcode + "'");
 			if (resultset.next()) {
 				city = resultset.getString("CityName");
 			}
@@ -245,26 +252,30 @@ public class Database {
 
 	public String register(String cpr, String email, String password, String name, String phone, String address,
 			Date date, String postcode) throws SQLException {
-		CallableStatement call = connection
-				.prepareCall("{call \"DTUGRP05\".UserRegister(?, ?, ?, ?, ?, ?, ?, ?, ?, ?) }");
-		call.setString("vCPRNo", cpr);
-		call.setString("vEmail", email);
-		call.setString("vPassword", password);
-		call.setString("vRoleID", "c");
-		call.setString("vFullName", name);
-		call.setString("vPhone", phone);
-		call.setString("vAddress", address);
-		call.setDate("vDateOfBirth", date);
-		call.setString("vPostcode", postcode);
-		call.registerOutParameter("vOutput", java.sql.Types.VARCHAR);
-		call.execute();
-		return call.getString("vOutput");
+		if (session != null && session.getAttribute("role").equals("e")) {
+			CallableStatement call = connection
+					.prepareCall("{call \"DTUGRP05\".UserRegister(?, ?, ?, ?, ?, ?, ?, ?, ?, ?) }");
+			call.setString("vCPRNo", cpr);
+			call.setString("vEmail", email);
+			call.setString("vPassword", password);
+			call.setString("vRoleID", "c");
+			call.setString("vFullName", name);
+			call.setString("vPhone", phone);
+			call.setString("vAddress", address);
+			call.setDate("vDateOfBirth", date);
+			call.setString("vPostcode", postcode);
+			call.registerOutParameter("vOutput", java.sql.Types.VARCHAR);
+			call.execute();
+			return call.getString("vOutput");
+		}
+		return loginString;
 	}
 
 	public User getUser(String cpr) {
 		User user = null;
 		try {
-			ResultSet resultset = statement.executeQuery("select * from \"DTUGRP05\".\"USERS\" WHERE \"CPRNo\" = '" + cpr + "'");
+			ResultSet resultset = statement
+					.executeQuery("select * from \"DTUGRP05\".\"USERS\" WHERE \"CPRNo\" = '" + cpr + "'");
 			if (resultset.next()) {
 				user = new User(this, cpr);
 				String email = resultset.getString("Email");
@@ -287,27 +298,34 @@ public class Database {
 		return user;
 	}
 
-	public String editUser(String cpr, String email, String password, String name, String address, String zipcode, Date date, String phone) throws SQLException {
-		CallableStatement call = connection.prepareCall("{call \"DTUGRP05\".EditUser(?, ?, ?, ?, ?, ?, ?, ?, ?) }");
-		call.setString("vCPRNo", cpr);
-		call.setString("vEmail", email);
-		call.setString("vPassword", password);
-		call.setString("vFullName", name);
-		call.setString("vAddress", address);
-		call.setString("vPostcode", zipcode);
-		call.setDate("vDateOfBirth", date);
-		call.setString("vPhone", phone);
-		call.registerOutParameter("vOutput", java.sql.Types.VARCHAR);
-		call.execute();
-		return call.getString("vOutput");
+	public String editUser(String cpr, String email, String password, String name, String address, String postcode,
+			Date date, String phone) throws SQLException {
+		if (session != null && session.getAttribute("role").equals("e")) {
+			CallableStatement call = connection.prepareCall("{call \"DTUGRP05\".EditUser(?, ?, ?, ?, ?, ?, ?, ?, ?) }");
+			call.setString("vCPRNo", cpr);
+			call.setString("vEmail", email);
+			call.setString("vPassword", password);
+			call.setString("vFullName", name);
+			call.setString("vAddress", address);
+			call.setString("vPostcode", postcode);
+			call.setDate("vDateOfBirth", date);
+			call.setString("vPhone", phone);
+			call.registerOutParameter("vOutput", java.sql.Types.VARCHAR);
+			call.execute();
+			return call.getString("vOutput");
+		}
+		return loginString;
 	}
 
 	public String deleteUser(String cpr) throws SQLException {
-		CallableStatement call = connection.prepareCall("{call \"DTUGRP05\".DeleteUser(?, ?) }");
-		call.setString("vCPRNo", cpr);
-		call.registerOutParameter("vOutput", java.sql.Types.VARCHAR);
-		call.execute();
-		return call.getString("vOutput");
+		if (session != null && session.getAttribute("role").equals("e")) {
+			CallableStatement call = connection.prepareCall("{call \"DTUGRP05\".DeleteUser(?, ?) }");
+			call.setString("vCPRNo", cpr);
+			call.registerOutParameter("vOutput", java.sql.Types.VARCHAR);
+			call.execute();
+			return call.getString("vOutput");
+		}
+		return loginString;
 	}
 
 	public String generateAccountID() throws SQLException {
